requirements.txt
beautifulsoup4==4.10.0
discord.py==1.7.3
python-dotenv==0.19.2
requests==2.26.0
lxml==4.6.4
PyDictionary==2.0.1
